Pulled AVAudioEngine extensions out to their own file
